requirements.txt
streamlit==1.28.1

PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

langchain==0.0.335
langchain-community==0.0.1

sentence-transformers==2.2.2
huggingface-hub==0.14.1
transformers==4.30.2

faiss-cpu==1.7.4
torch==2.0.1
accelerate==0.21.0